Fixed tests and formatting

+++ solutions/max_integer.py
@@ -14,16 +14,16 @@ Author: Tamir Elwaleeed
 def max_integer(my_list: list) -> int:
     """Returns the max integer in a list of integers
         If the list is empty, the function returns None
-        
+
     Parameters:
         my_list: list, the list to find the max from
-        
+
     Returns -> int: the max integer from the list
-    
+
     Raises:
         AssertionError: if input is not a list
         AssertionError: if one of the elements isn't a number
-    
+
     Examples:
         >>> max_integer([1, 2, 3, 4, 5])
         5
@@ -31,11 +31,11 @@ def max_integer(my_list: list) -> int:
         -1
         >>> max_integer([])
         None
-    """ 
+    """
     assert isinstance(my_list, list), "input must be a list"
     for i in my_list:
-        assert isinstance(i, int), "elements must be numbers"
-        
+        assert isinstance(i, float), "elements must be numbers"
+
     if len(my_list) == 0:
         return None
     result = my_list[0]
